added support for updating transaction split

+++ src/models/response.rs
@@ -1,7 +1,7 @@
 //! response
 //! ========
 //! Holds the generic response templates for the API
-use crate::utils::string_to_u8;
+use crate::utils::string_or_number_to_u16;
 use serde::{Deserialize, Serialize};
 
 /// Generic response body template for the API
@@ -22,16 +22,20 @@ pub struct Response<T> {
 #[serde(rename_all = "camelCase")]
 pub struct Meta {
     /// This is the total number of transactions that were performed by the customer.
-    pub total: u8,
+    #[serde(deserialize_with = "string_or_number_to_u16")]
+    pub total: u16,
     /// This is the number of records skipped before the first record in the array returned.
-    pub skipped: u8,
+    #[serde(deserialize_with = "string_or_number_to_u16")]
+    pub skipped: u16,
     /// This is the maximum number of records that will be returned per request.
-    #[serde(deserialize_with = "string_to_u8")]
-    pub per_page: u8,
+    #[serde(deserialize_with = "string_or_number_to_u16")]
+    pub per_page: u16,
     /// This is the current page being returned.
-    pub page: u8,
+    #[serde(deserialize_with = "string_or_number_to_u16")]
+    pub page: u16,
     /// This is how many pages in total are available for retrieval considering the maximum records per page specified.
-    pub page_count: u8,
+    #[serde(deserialize_with = "string_or_number_to_u16")]
+    pub page_count: u16,
 }
 
 /// This struct represents the authorization data of the transaction status response
